+++ tests/legs.rs
@@ -5,7 +5,6 @@
 
 use hashbrown::HashSet;
 use metropolis::agent::{agent_index, Agent};
-use metropolis::learning::{ExponentialLearningModel, LearningModel};
 use metropolis::mode::trip::event::RoadEvent;
 use metropolis::mode::trip::results::{LegResults, LegTypeResults, RoadLegResults, TripResults};
 use metropolis::mode::trip::{DepartureTimeModel, Leg, LegType, RoadLeg, TravelingMode};
@@ -140,15 +139,12 @@ fn get_simulation() -> Simulation<f64> {
     let network_parameters = NetworkParameters {
         road_network: Some(RoadNetworkParameters::from_recording_interval(Time(5.0))),
     };
-    let parameters = Parameters::new(
-        Interval([Time(0.0), Time(50.0)]),
-        1,
-        network_parameters,
-        LearningModel::Exponential(ExponentialLearningModel::new(0.0)),
-        vec![StopCriterion::MaxIteration(1)],
-        1.0,
-        None,
-    );
+    let parameters = Parameters {
+        period: Interval([Time(0.0), Time(50.0)]),
+        network: network_parameters,
+        stopping_criteria: vec![StopCriterion::MaxIteration(1)],
+        ..Default::default()
+    };
 
     Simulation::new(agents, network, parameters)
 }